set up new files structure
created new classes: DbUtils, ScoreService and UserService
DbUtils contains one method: sendResponse
UserService and ScoreService contain static classes previously present
in DbController class
