+++ app/src/main/java/org/example/database/DbUtils.java
@@ -1,0 +1,17 @@
+package org.example.database;
+
+import com.sun.net.httpserver.HttpExchange;
+
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.charset.StandardCharsets;
+
+public class DbUtils {
+    public static void sendResponse(HttpExchange exchange, int statusCode, String message) throws IOException {
+        byte[] response = message.getBytes(StandardCharsets.UTF_8);
+        exchange.sendResponseHeaders(statusCode, response.length);
+        try (OutputStream os = exchange.getResponseBody()) {
+            os.write(response);
+        }
+    }
+}
